Update patch set 1
Patch Set 1:

yeah, and I guess that just bails out with a exit(1) ... same as if some internal issue that would not resolve itself if you did it N times.

we go to some effort to get pypi mirrors close to the tests, so i don't recall ever seeing this issue in the openstack gate.  my point being that i think future pip failures are mostly likely to be incorrectly retrying with this change, at least in the gate.

there is some precedent for checking the output, perhaps for the "timed out --" bit ... we do a similar thing for yum which doesn't give useful errror codes to distinguish errors.

Patch-set: 1