Update patch set 2
Abandoned

I've cooled on this idea.  This was really trying to work-around using "sudo" on the call, but as noted it doesn't work if you're trying to create the file -- the thing the "api" (such as it is) is trying to do.  i don't think complicating this further helps

See Iaceb8d42ce37be728adae6fd0a30a1f9d33d4029

Patch-set: 2
Status: abandoned