Update patch set 1
Patch Set 1: Code-Review-1

It is actually sourcing stackrc where the problem occurs and skipping that is hard.  Those function calls ideally should not be in stackrc but HOST_IP is needed in a number of places where that position is required.

Please add to the comment the reasons for doing this (working around get_default_host_ip() in stackrc) and the caveat of the IP address falling in the configured FIXED_RANGE.

Also, stackrc does have a note about the overlap of IP address with FIXED_RANGE.  One could argue that this is a configuration problem in trying to use a FIXED_RANGE that is not in fact unused in the operating environment.

Patch-set: 1
Label: Code-Review=-1