Update patch set 2
Patch Set 2: Code-Review+1

I remember having a version of Neutron from maybe .... Folsom era? and we did a migration following the documentation from the OVS plugin to ML2 - and we got bit hard by latin1 vs. UTF8 bugs in the migration. So, this is a big problem - I don't know if fixing in DevStack is enough - but maybe this is a step in the right direction? Not sure.

Patch-set: 2
Reviewer: Gerrit User 4656 <4656@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1