Update patch set 1
Patch Set 1:

So, I'm not sure how to proceed here. The neutron-vpnaas repo has a devstack plugin now, and people can specify to enable the plugin when using devstack.

This commit is attempted to remove the VPN setup from the DevStack repo, as it is no longer needed. However, many of the Jenkins jobs in the DevStack repo are running DevStack, but the VPN plugin is not being specified, so those tests are failing.

Do we add the enable_plugin line for VPN for all the failing tests, all dsvm based tests, all voting tests, or do we somehow eliminate VPN from the tests?

Patch-set: 1