Support new ironic ENROLL state
Starting with version 1.11 of the bare metal API nodes appear first
in ENROLL state, not AVAILABLE. This change detects support for 1.11
API version and explicitly requests it on node-create.

Thus we'll avoid breakage if API version >= 1.11 will ever become
the default in ironicclient.

Nice side effect of this patch is that power credentials for nodes
will be verified as part of node enrollment.

Change-Id: I0e33a49a6c3fdc092a62d4de1b0f12f462312f3e