Update patch set 3
Patch Set 3:

Ok Ryan I must have made a mistake somewhere (which is very possible). Still I am a bit afraid that this could lead to a bug or a weird situation that would be hard to debug. But I am a bit "young" in that area so I'll let other reviewers review :)

Patch-set: 3