Move horizon apache root to /dashboard
With keystone's move to /identity, a conflict in for resources was
created as both keystone and horizon used /identity. The keystone
config took precedence and rendered API output in the horizon UI.

This patch sets the root for horizon to /dashboard and serves all
horizon content from there. Additionally, a RedirectMatch has been added
to the apache config for horizon to redirect '/' to '/dashboard' this
will allow the implementation to change without being immediately
painful to users.

There is one patch in horizon that needs to land before this change can
happen seemlessly, noted below.

Closes-Bug: #1478306
Depends-On: I280969682dccc4efc2252771ef085bd0a0a1acfa

Change-Id: I6fbca5cea9e44df160afbccc71bd045437657320