Update patch set 2
Patch Set 2:

> Hirofumi Ichihara
> I guess that we don't need metadata_proxy_shared_secret variable. If users need it, they can set into local.conf.

But, "If users need it, they can set into local.conf", this is not the case as devstack doesn't provide a way to set this config w/o this patch. 

Are you suggesting to make it optional so it sets it only when you configure local.conf?

Patch-set: 2
Label: Verified=0