Update patch set 1
Patch Set 1: Code-Review-1

I agree, this is working in the gate so i assume you've got an old OSC that doesn't support -f value for list. I don't know when this was added. 

If there is some reason to support older versions of OSC then i'd be ok with reverting to the get_value methods.

Patch-set: 1
Label: Code-Review=-1