Add check for existing plugin
If you call enable_plugin twice with the same name, the details are
updated correctly, however the plugin name gets added a second time to
the list.  This causes the plugin code to be executed twice and you'll
end up with two sets of processes running.

A check was added to enable_plugin so that the name only gets added
to the list if it doesn't already exist.

Change-Id: Ifabfda444aed65467788b5e46a0cafe2a3b7883d