Update patch set 3
Patch Set 3:

I see that even if it's blank, that's still treated as a secret and an hmac signature is generated based on that. So I'm not sure it gains us much other than rotating the key.

However, it does have a downside. Since it's randomly generated, a multi-node devstack setup will now break since each run will produce different keys.

With that in mind, I don't think it's a good idea to go forward with this.

Patch-set: 3