Copy rootwrap filters
This patch copies os-brick's list of rootwrap filters for commands
that are needed to execute from /usr/local/etc/os-brick/rootwrap.d
to service's rootwrap.d directory to enable filters.

This patch has the following dependencies which are all merged now:

The os-brick patch I2b1e657b87c7b27548200a20b991f34c3413c24b adds
os-brick's list of rootwrap filters for commands that are needed
to execute. That patch is merged and it is in os-brick 0.4.0
which is already released.

This devstack patch also needs the minimum required version of
os-brick in global-requirements to be bumped to 0.4.0. That change
I2f5ca36da2e5d84be260e32da269802993d7722f is merged as well.

Partial-Bug: #1479842

Change-Id: I9d77783c4cc8d8db7fff21543526fd4060f8b12b