Ensure a service stop during stop_process
The stop_process is called from grenade test cases, but
sometimes this method can't stop process properly and
it causes gate test error like this.
https://bugs.launchpad.net/oslo-incubator/+bug/1446583

This patch adds a feature to wait for a service to stop and
sends SIGKILL to forsibly kill the service.

The wait is up to 10 seconds because sometimes a service
takes a while to shutdown. This logic is already implemented
grenade/inc/upgrade, this pach ported the logic into
devstack script.

Also some components have a feature to shutdown their service
gracefully. In this case, the service might not stop within
10 seconds, therefore this patch also adds sending SIGKILL
to forcibly kill the service again.

Change-Id: Ib171fa15c2912073601057e5e3e02c558f62d832