Update patch set 4
Patch Set 4: Code-Review-1

DevStack has documentation about how much RAM a system needs. However, up until earlier this year it was incredibly out of date, I think it was either 512MB or 1GB of ram, which was really what DevStack needed ... in 2012. I think that the velocity of OpenStack is such that it's difficult to keep a good set of recommendations as to the minimum amount of resources that are required for a successful stack.sh run. 

So, it's not a great reason to -1 this, but we've done a bad job at documenting what the minimums are, and that was just docs, not code (which is even less flexible and forgiving)

Patch-set: 4
Reviewer: Gerrit User 4656 <4656@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1