Update patch set 3
Patch Set 3:

(12:52:10 PM) jeblair: mriedem: aiui, os-brick is installed from pypi releases only in devstack, so it is not possible to test potential changes to os-brick with other software.

(12:52:16 PM) jeblair: mriedem: sdague may be able to elaborate

--

Looking at devstack, yes it either pulls os-brick from git if you've configured the job to do so, else it pulls the dependency from pypi, which is what's happening here.

So maybe we need to make the os-brick change depend on the devstack change and then while the devstack change will fail, we'll verify if things are working by checking the results on the os-brick change.

Patch-set: 3
Label: Verified=0