Update patch set 1
Patch Set 1:

I sure can use PUBLIC_INTERFACE, but the name is confusing for my purpose because I intend to use the interface for tenant network instead of public network.
Could you explain the purpose of part:
# Configure bridge manually with physical interface as port for multi-node
_neutron_ovs_base_add_bridge $OVS_PHYSICAL_BRIDGE

if you don't add port into bridge, what is the use of it?

Moreover, be GUEST_INTERFACE_DEFAULT for xenserver, the code in lib/nova will overwrite the variable to the first interface rather than the value set in local.conf.

Patch-set: 1