Update patch set 2
Patch Set 2:

Jordan,

can you enlighten me why it should?

i did this way to avoid breaking jobs.
it isn't realistic (for me) to audit every CIs and jobs
which is using tempest.
besides, this is not the only option which is false by default.
for example, run_validation is false by default.

Patch-set: 2