Update patch set 2
Patch Set 2:

>i did this way to avoid breaking jobs
No, you did not. You have submitted a new job in Tempest ( https://review.openstack.org/#/c/192059 ) which is enabled and running by default... (as far as I understood)

Now if this patch is merged, and lib/tempest is changed, it means that the Tempest test "verify_metadata in tempest/scenario/test_server_basic_ops.py" will never be run by the Gate. And this not what the Tempest team wants.

Patch-set: 2