Update patch set 2
Patch Set 2:

Jordan,

well, it seems my memory is a bit dim.

you'are right the option in tempest is true by default.
on the other hand, it's ignored unless run_validation=True,
which is false by default.

i guess it's fair to expect maintainers of CIs/jobs to tweak
this option when they turn run_validation true.
in that sense making this true by default should be ok.

i still don't understand why you think this will never be run
on gate if it's false by default, though.

Patch-set: 2