Update patch set 4
Patch Set 4:

One reason is that we need the integration test infra for kafka to finish a BP for oslo_messaging:  https://review.openstack.org/#/c/189003/
(it is required to support for integration test when we add new oslo_messaging drivers.) 

Second reason is thet kafka is the same abstraction layer with rabbitmq, zeromq, and other messaging brokers whose codes seem to be managed in devstack repo.

Actually I'm not sure the code is needed to be in devstack repo (our main purpose is making the integration test infra), so if you think it's better to be finish this aims by devstack plugin, we'll do by that way.

Thanks for the review.

Patch-set: 4
Reviewer: Gerrit User 12850 <12850@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0