Workaround large-ops wedge with uncapped novaclient in osc 1.0.1
python-openstackclient 1.0.1 pulls in uncapped novaclient which breaks
the caps for novaclient in other projects in Juno.  This only happens in
the large-ops job because cinder isn't installed and therefore cinder
doesn't install capped novaclient per juno global-requirements for osc
is installed.  We can't fix osc and bump it to 1.0.2 since osc is capped
at <=1.0.1 in global-requirements for stable/juno, so we're stuck.

So the workaround is to install python-novaclient from
global-requirements prior to installing python-openstackclient but only
in the case of the large-ops job, which we identify from the usage of
the fake virt driver.

Closes-Bug: #1482350

Change-Id: Id8cc112800d757bfd509cc48087b769410356920