Provide alternate path for neutron plugin config files
Since we are in the process of decomposing vendor specific plugins,
some plugin config files may no longer exist in neutron repo.
We need to provide an alternate path for them. So neutron-leagcy
can still access them if not found in the previous location.

The new path is defined in Q_PLUGIN_ALT_CONF_PATH.

Partial-bug: #1483453
Change-Id: I2fbe8b4ce5a74164a4770a58a7636b9bb3dc7d3b