Allow installing python-pip from the package
When python-pip is installed by https://bootstrap.pypa.io/get-pip.py
into CentOS7, there is a case in which it does not move.

The python-pip, which is provided from the distributor or the like, it
can be used more stable. By upgrading to the latest version by the
capability provided in itself, it becomes possible to use the latest
version more stably.

Therefore, install_pip.sh is modified so that you can install python-pip
from the package provided by the distribution and so on.

Change-Id: Iae94a205cedbd0ca4030a16244e8dfc1612ee0e3
Closes-Bug: #1501284