Allow installing python-pip from the package provided by the Distribution.
As long as it does not require recent modify, the python-pip,
which is provided from the distributor or the like, it can be
used more stable. At least, it does not cause problems even when
the latest development version had temporarily degraded.

Update of python-pip are repeated sufficiently, then, mature
package is provided by the distributor or the like. In most
cases, it suffices in this package.

However, in case the recent modify become necessary, the latest
version provided by the community will be installed when the
package provided by the distributor or the like is lower than
specified version. Now, version 1.4.1 is specified.

Change-Id: Iae94a205cedbd0ca4030a16244e8dfc1612ee0e3