Update patch set 1
Patch Set 1:

In fact the problem that I have faced in mid-June this year, when I installed the DevStack to CentOS7, the pip command could not return the results of '--version', then stack.sh is failed. I uninstall the pip of the latest version, re-install the python-pip package of epel, and I commented out where to install the latest version of pip, then OpenStack has started without any problems.

Any OSS that does not limit the targeted distributions, in all of the distributions (Ubuntu, RHEL, Fedora, etc.), it is difficult to prevent all the degraded that depend on distributions with the latest development version. Certainly, I will guess that you have often had issues with older versions of the pip.
However, it passed more than a dozen months, than the pip to have a proven as part of each distribution but is a little older than the latest version, the latest development version there is a risk of degraded, the more risk to have issues. In fact, the update frequency of DevStack recently became not often. It does not look like DevStack has been modified to use the modifications made to the latest pip.

As other apps to use pip that installed package of distributions, if there are no special circumstances, we'll have issues less than the latest development version, especially cover the various distributions. Of course, since it is necessary to minimum required functions to pip is implemented with, and I added the function to install the latest version if it is lower than specified version.
However, here if you specify the version of the pip, because it also becomes directly DevStack requirements, versatility is higher if you specify an older version as much as possible.
If DevStack will increase targeted distributions, it is more better to use the pip package for distributions.

Patch-set: 1
Label: Verified=0