Support BlockDeviceDriver as as cinder backend
BlockDeviceDriver as an option, that available in Cinder,
but it's unsupported by devstack. The aim of this patch
to introduce BlockDeviceDriver at devstack.

This cinder backend is going to be used by Sahara in testing, 
because it can improve I/O performance, which is important for
Hadoop performance. More info about BlockDeviceDriver can be
found at reference link.

Reference link:
https://wiki.openstack.org/wiki/BlockDeviceDriver

Change-Id: Iff0ee100750e70cb5c93110aaf8595675bc2556d