Update patch set 2
Patch Set 2:

Yes! using the service catalog for every service to service communication would be fantastic. I fight this on just about every service. 

For example why do i still need to tell nova the URL for neutron? Or glance to swift? Or everything to do wiht ironic.

This is a great goal but it's way outside of what we can fix quickly in devstack.

Patch-set: 2
Label: Verified=0