Update patch set 3
Patch Set 3: Code-Review-1 -Workflow

So I just realized that this will incorrectly set the option to true even if the metadata service isn't in use. For example we run with just config drive on most of the jobs but this is still setting the option to true. I think we should check the devstack env to check if we're configuring and running with the metadata service and use that instead of a new variable which is unused anywhere

Patch-set: 3