Update patch set 5
Patch Set 5:

Hi Ian, no this isn't WIP. It's ready to go.

Ie8a9933e41cc0bfc26ac68e1338c762f5a982521 had been widely disagreed with and is therefore WIP until it can be discussed in detail and a new solution agreed upon. This devstack patch is needed to workaround the problem until then - or possibly forever if the existing functionality is agreed to be desired. Since Ie8a9933e41cc0bfc26ac68e1338c762f5a982521 is likely never to merge I'm not sure it would be very useful to reference it in the commit.

The bug report captures the details of why this patch is needed, but I can copy some of that detail into the commit message if you prefer. Thanks.

Patch-set: 5
Label: Verified=0