Update patch set 5
Patch Set 5:

I think adding the ability to configure this in devstack is fine; however I am still proposing that this driver be removed from Cinder.  It serves no real purpose to anybody but the Sahara team.  It's not tested, it's not well maintained and it doesn't meet the majority of the required features that other drivers are required to have.

If there isn't a CI running against it in M (and in it's current state it can't run the tempest tests) I will propose removal yet again.

Patch-set: 5
Label: Verified=0