Update patch set 6
Patch Set 6:

Now that https://review.openstack.org/#/c/207553/ is merged you could propose a version bump for os-brick (a minor version bump, so 0.4.0) and then once that's out you'll probably need to raise the minimum required os-brick in global-requirements for master to be <=0.4.0 so that we're sure it's used here and in nova.  Keep in mind that dep freeze is next week, so you probably want to get this done by 9/1.

Patch-set: 6
Label: Verified=0