Specify regexp instead of listing libraries
We should be able to specify:
LIBS_FROM_GIT=oslo.*

for example to pick up all oslo libraries from git, with earlier
code we would have had to list each oslo library in LIBS_FROM_GIT
and keep updating this list when we develop new oslo libraries.

Note that we keep the old functionality (list libraries) and allow
regexps as well.

We need this to be able to run an experimental job in Nova to
run dsvm tests against the master of all oslo libraries (instead of
picking them up from pypi mirrors)

Change-Id: I1f505e81c5791fb823ea84032334bed33cad8c6d