Update patch set 16
Patch Set 16: Code-Review+1

I am not neutron infra liaison, but I am quite familiar with devstack-gate, tempest, and neutron. This change does not impact any gate job as the variable used here it is not set in devstack_gate.

If a plugin library sets it, then it uses this path to copy configuration files for the plugin from the appropriate directory. It's simple as that. with a little tweak concerning whether the path is absolute or not.

The only thing on which devstack-core input would be welcome are:
- whether it is ok to let the devstack plugin only specify where to copy the files, and then let devstack copy them, or whether we should remove the ability of doing so from neutron-legacy
- since this change affects neutron-legacy I wonder whether this should a functionality available only to 'legacy' plugins. Decomposed plugins should do everything in the devstack plugin. In that case we need a solution as devstack fails to copy configuration files for decomposed neutron plugins.

Patch-set: 16
Reviewer: Gerrit User 261 <261@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1