Update patch set 16
Patch Set 16:

I'd like to mention again (it's been a while) that I am really not a fan of the way the config values are handled WRT generally not using a leading '/' so it is required in the code.  This is the _only_ place in DevStack that does this and it can get confusing.  My intention is still to remove that behaviour in the neutron rewrite, if I ever get back to it...

Patch-set: 16
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0