Update patch set 3
Patch Set 3:

There is an instance delete that fails starting around here when we start tearing down the container:

http://logs.openstack.org/48/219448/3/check/gate-tempest-dsvm-full/672e11a/logs/screen-n-cpu.txt.gz#_2015-09-03_22_07_10_171

2015-09-03 22:07:10.171 DEBUG nova.virt.libvirt.driver [req-32dc93c1-a619-4311-83dd-e347855decd2 tempest-ServersAdminNegativeTestJSON-105603620 tempest-ServersAdminNegativeTestJSON-414787637] [instance: 93dce47b-ad5e-47d8-9c9c-a4047d094f0f] Attempting to teardown container at path /opt/stack/data/nova/instances/93dce47b-ad5e-47d8-9c9c-a4047d094f0f/rootfs with root device: /dev/nbd8 _teardown_container /opt/stack/new/nova/nova/virt/libvirt/driver.py:691
2015-09-03 22:07:10.172 DEBUG nova.virt.disk.api [req-32dc93c1-a619-4311-83dd-e347855decd2 tempest-ServersAdminNegativeTestJSON-105603620 tempest-ServersAdminNegativeTestJSON-414787637] No device found for path: /opt/stack/data/nova/instances/93dce47b-ad5e-47d8-9c9c-a4047d094f0f/rootfs __init__ /opt/stack/new/nova/nova/virt/disk/api.py:311
2015-09-03 22:07:10.172 DEBUG nova.virt.disk.api [req-32dc93c1-a619-4311-83dd-e347855decd2 tempest-ServersAdminNegativeTestJSON-105603620 tempest-ServersAdminNegativeTestJSON-414787637] Release nbd device /dev/nbd8 teardown_container /opt/stack/new/nova/nova/virt/disk/api.py:476
2015-09-03 22:07:10.173 DEBUG oslo_concurrency.processutils [req-32dc93c1-a619-4311-83dd-e347855decd2 tempest-ServersAdminNegativeTestJSON-105603620 tempest-ServersAdminNegativeTestJSON-414787637] Running cmd (subprocess): sudo nova-rootwrap /etc/nova/rootwrap.conf qemu-nbd -d /dev/nbd8 execute /usr/local/lib/python2.7/dist-packages/oslo_concurrency/processutils.py:230

It doesn't happen in the disk API, but we do call qemu-nbd -d on the nbd device.

And then the instance file delete fails here:

http://logs.openstack.org/48/219448/3/check/gate-tempest-dsvm-full/672e11a/logs/screen-n-cpu.txt.gz#_2015-09-03_22_07_11_085

I'm wondering if qemu-nbd -d is asynchronous (it probably is, most things in qemu/libvirt are asynchronous).

So we could (1) add some code to see if the nbd device is really freed up after calling qemu-nbd -d, but that might race with another thread spawning an instance that gets that device immediately after it's freed up, or (2) we add some retry logic around the instance file delete code if we get an EBUSY error (with maybe some lsof logging for debug).

Patch-set: 3
Label: Verified=0