Update patch set 7
Patch Set 7:

Matt, this may be just a case of a horrible brainstorming idea, but what if os-brick had something like a get_rootwrap_filters() call that would return the contents of its filters. Then any consumers on startup could be responsible for calling that and writing out the current filters to its own rootwrap.d location.

This would help cases in an existing deployment where a new version of os-brick is released. On service restart it would just pull in the new filters without requiring the admin to be aware that os-brick was necessarily updated and required a manual copy of a (somewhat hard to locate) file.

Feel free to point out all the holes in this idea. :)

Patch-set: 7