Update patch set 7
Patch Set 7:

Sean, not a terrible idea.  I'm not entirely sure if the user nova is running under is going to have write access to /etc/nova/rootwrap.d, that again seems like a config/deploy management issue, but it seems like it could work.  I think we're open to any ideas here.  Could also be posted to the ML (dev and operators lists) on impacts/heads-up/ideas here.

Patch-set: 7