Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

some queries before we go ahead

I don't see any mention of the existing use of SWIFT_DEFAULT_BIND_PORT in lib/ironic?  That uses HOST_IP rather than SERVICE_HOST (and defaults to http, part of what this change fixes)

are we sure that this is the right thing to do wrt to that variable?

maybe the url fixups (service protocol, etc) should be split out of this change, and merged in with any fixes to the ironic urls, then have this follow on from that

Patch-set: 1
Label: Code-Review=-1