Update patch set 17
Patch Set 17:

OK. That's the part I don't understand.

In our plugin, the old config file was in
neutron:/etc/neutron/plugins/vmware/.
The new config is in vmware-nsx:vmware_nsx/etc/.
So either I define Q_PLUGIN_CONF_FILE as the former or the latter. The line will break.

  cp $NEUTRON_DIR/$Q_PLUGIN_CONF_FILE /$Q_PLUGIN_CONF_FILE

Is that right?

Patch-set: 17