Update patch set 17
Patch Set 17:

Hi Shih-Hao Li, what you really want is to get rid of all the vmware related code from [1,2] and centralize it in [3]. With the use of extras.d and external plugin support in Devstack, this code has no longer reason to exist in the devstack tree. If you do that, and follow the OVN use case I showed you, you can successfully integrate your plugin with Devstack and avoid fiddling with a logic that is labelled as 'legacy'.

[1] https://github.com/openstack-dev/devstack/tree/master/lib/neutron_plugins
[2] https://github.com/openstack-dev/devstack/tree/master/lib/neutron_thirdparty
[3] https://github.com/openstack/vmware-nsx/tree/master/devstack

Patch-set: 17