Update patch set 17
Patch Set 17:

@Armando - the OVN uses the ML2 etc file which is still in tree. If you decompose that then that too will break.
What this patch essentially does, which I am sure that you guys understand, is just enable us to specify the path of the decomposed etc file.
We can work around this by using the same neutron conf file and just updating that.
But that will kind of break the broken model.

I am not sure if you have tried this with the other external plugins, but all of those that use the upstream bits require this. There is/was a similar plumgrid one too.

Patch-set: 17