Update patch set 7
Patch Set 7: Code-Review+1

This patch needs to land to get the rootwrap filters updated and installed for Nova and Cinder.  

We are being forced to patch Nova and Cinder in the mean time, but that's not the right way to go.

Patch-set: 7
Reviewer: Gerrit User 5997 <5997@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1