Update patch set 7
Patch Set 7:

We should probably abandon this for now, we're moving forward with tactical fixes for liberty to just make sure the os-brick.filters contents are in volume.filters in cinder and compute.filters in nova.  We'll use bug 1479842 to track changes to oslo.rootwrap for mitaka to make that more dynamic.

Patch-set: 7