Update patch set 3
Patch Set 3:

@Chmouel: well, fixing it in swiftclient would solve this issue, but might make more problems in the future. In fact we would need to modify OS_AUTH_URL in swiftclient then, and that might not be wanted by every user.

Also, the real problem in this case is keystoneclient; if an OS_AUTH_URL without a trailing version identifier is used, it fails (for example doing a "keystone token-get"). To me it seems that this is the correct behavior, and the failure is due to an incorrect usage (which would be fixed by this patch).

Patch-set: 3
Reviewer: Gerrit User 6968 <6968@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0