Install database python client if a db is configured in local.conf
If `DATABASE_TYPE` is configured in `local.conf`, the database backend
is currently configured with `initialize_database_backends` even if no
database backend is enabled.

On a multi-nodes Devstack environment, such as devstack-vagrant, the
compute node currently fails because it does not have PyMysql. This
compute node has no database backend enabled, but has to connect to
the database on another node.

We should install the python client if DATABASE_TYPE is set, even
if no database backend is enabled.

Change-Id: Iffd5f7243a0dfdbe56cf6b9a87b96ed7678c81dd