Configure tempest to not use security groups for cells
Currently, any tests which use security groups are excluded from
the gate-tempest-dsvm-cells job since cells does not support
security groups.  The exclusion is done using a blacklist
in:
http://git.openstack.org/cgit/openstack/nova/tree/devstack/tempest-dsvm-cells-rc

Change devstack to set the tempest config option,
compute-feature-enabled.security_groups,
to False if cells are enabled.  This will allow the blacklist
to be pruned.

Another change set is dependent on this change:
Nova change to remove the security group related tests
from the blacklist.
https://review.openstack.org/#/c/226043/

Change-Id: I4f74b80024a9bcc374d374d86052dceb40598d32
Closes-Bug: 1495653
Depends-On: Icd62f42809380de6d638b343d3322b102b0d1d1f