Disable security group tests in tempest-dsvm-cells-rc using a config option
Currently, the tempest security group tests for tempest-dsvm-cells-rc
are blocked becausecells doesn't support security groups.  The block is
now done using a nova whitelist:
http://git.openstack.org/cgit/openstack/nova/tree/devstack/tempest-dsvm-cells-rc.
Change to block the tests based on a new tempest config option -
compute-feature-enabled.security_groups.

Also, some of the tempest scenario tests attempt to create a server with a
security group.  These will not work with cells, so use the new config option
to block those tests.

This commit changes devstack to set the compute-feature-enabled.security_groups
option to False if cells are enabled.

Change-Id: I4f74b80024a9bcc374d374d86052dceb40598d32
Partial-Bug: 1495653
Depends-On: Icd62f42809380de6d638b343d3322b102b0d1d1f