Update patch set 3
Patch Set 3: Code-Review-1

-1 because we should include even a minimal test-case (tests/test_functions.sh probably)

This is a fun hack ... but putting in a gate-breaking time-bomb seems like, well, not a good idea.  

By Murphy's law, it will happen at a bad time and cause unintended flow-on affects.  Although you can argue that we (people related to upstream infra) can fix the gate quickly by pushing through a change, there's plenty of external users that don't update every second and have other "interesting" setups; we should attempt to minimise their pain.

Putting out a warning in the middle of thousands of lines of devstack log file is as good as doing nothing, I admit.

Maybe a compromise is to move the log messages for timed-out work-arounds into a separate log file?  This way they're easy to find, but we're not going to shoot ourselves in the foot.

Patch-set: 3
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1