Support for oslo.messaging pika driver for RabbitMQ
Oslo team is writing a new driver for RabbitMQ using
the pika library. We'll need to support switching
between the existing kombu library and the new pika
library based as well to be able to develop and
trouble shoot the new driver. This patch makes it
configurable.

Change-Id: Ie3cb4c2c9868d0c859f1a8dccde44df6f69ceb81
Depends-On: I7bda78820e657b1e97bf888d4065a917eb317cfb
Implements: blueprint configurable-rpc-backend
Implements: blueprint rabbit-pika