Update patch set 2
Patch Set 2:

So Monty's reasoning for moving that up in the first place was to aid debugging failed runs.  Is it possible to create the files early and re-write them late to take in to account plugins and the remaining services?

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0