Change output filed number from "nova floating-ip-create" for IP
The existing floating_ips.sh script fails when run on a freshly
installed devstack with an error similar to the following:

[Call Trace]
./floating_ips.sh:158:die
[ERROR] ./floating_ips.sh:158 Failure adding floating IP 1 to ex-float

It appears that the code expected the IP address in the first field,
whereas "nova floating-ip-create" returns an ID in the first field and
an IP in the second field. The simplest fix seemed to be to have the
field parser choose the correct field based on current nova output.

This patch changes "get_field 1" to "get_field 2" for the two
"nova floating-ip-create" commands.

Change-Id: Ibcbb06fbd761b88a4869fc583d23d96cc9debc81