Update patch set 2
Patch Set 2: Code-Review-1

the point of the nano instance was to do really small cirros images. I feel like we're now at the point where the power specific instances are as big as default flavors we ship, so instead of making special nano for these, we should instead make it so on power you just specify different flavors that are the defaults.

Patch-set: 2
Label: Code-Review=-1