Double API_WORKERS for nova-conductor when running large ops
Commit 1ce19ab76d67a89b04f907f1d292d013a3b699e0 dropped API_WORKERS from
nproc/2 to nproc/4 and also started using API_WORKERS for the number of
conductor workers, so in gate runs that dropped conductor workers from 8
to 2.  We're now seeing instance build timeouts in the large ops job.

This change doubles the number of conductor workers so that in a gate
run we have 4 conductor workers and keep API workers at 2.

We could just do nproc/4 for the large ops jobs, but it seems nova-api
isn't the problem.  Plus the change gets messier in devstack when you
have to take into account the function that sets the workers when
running nova with Apache.

Closes-Bug: #1500615

Change-Id: Ie6ef855fce0a99c930d479b7459c15b69e8de499