Update patch set 1
Patch Set 1: Code-Review+1

I like this because it keeps us running with only two api workers, which tests what we want: the minimum number of workers to pass the tests reliably. If this isn't enough and we had to bump api workers back up, that would be a data point about a different thing, which would be useful.

Patch-set: 1
Label: Code-Review=+1