Update patch set 1
Patch Set 1: Code-Review+2 Workflow+1

I'm happy with getting this in here, but it would be really great if we could also get this to be the defaults in swift as well. Divergence from the projects is something we try to avoid if we can.

Patch-set: 1
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=+1