Update patch set 3
Patch Set 3: Code-Review-1

Please, give us enough context to review this without going on an archaeological dig

As I believe I've found out, the reason behind this is to use "pika" as opposed to "rabbit"/"kombu" because the latter is less maintained.

Please link to [1] and explain this in the commit.  I also think this need some depends-on: on the outstanding changes, because it's not something we want without those committed.  also wants a note in doc/source/configuration where we talk about rpc backends

So given the context of this, and our general unwillingness to re-introduce too many variables in to the RPC path after going to some effort to reduce it in  I8073a895c03ec927a2598eff6c2f01e5c82606fc, I'd prefer to see us checking a externally set value is either rabbit or, i guess, "pika", and throwing an error if it is anything else.  that way we are explict in what we support.

[1] https://blueprints.launchpad.net/oslo.messaging/+spec/rabbit-pika

Patch-set: 3
Label: Code-Review=-1