Update patch set 4
Patch Set 4: -Code-Review

It is still a bit confusion, the s3 endpoint creation does not means the server is running.

The s3 service can be:
- n-obj  (default)
- swift with swift3 plugin
- whatever thing with s3

The ec2 is usually n-api.

BTW: ALL endpoints should be created as soon as keystone is up, and not waiting for individual service install script.

Patch-set: 4
Label: Code-Review=0