Update patch set 5
Patch Set 5:

The reason nova-compute is still considered running is that n-cpu.pid contains the pid of sg that is used to start nova-compute, not of nova-compute process itself. So if we send kill <n-cpu.pid>, nova-compute process won't be killed because sg doesn't pass signals to child processes.

(Here [1] you can see that nova-compute is actually running:

2015-10-06 12:08:07.823 | stack    26555  3.5  1.5 1843432 117024 ?      Sl   11:54   0:29 /usr/bin/python /usr/local/bin/nova-compute --config-file /etc/nova/nova.conf

)

That is why I modified the code that is writing to .pid file to put a correct pid there.

[1] http://logs.openstack.org/64/208064/6/check/gate-grenade-dsvm/2d6bac0/logs/grenade.sh.txt.gz#_2015-10-06_12_08_07_823

Patch-set: 5
Label: Verified=0