collect service failures even when not under screen
When not running under screen we end up potentially missing some
services not starting because we don't check that they actually
started and write out error logs.

We mitigate this by ensuring that pids we believe should be running, are at
the end of the run.  This does potentially have a race on pid wrapping, but
it is a 95% solution that will help us identify some breaks.

Change-Id: I4dc7b956024fb03f4db59b0a7654ae5cd12bd206
Closes-Bug: #1504118